requirements.txt
rpaframework==12.4.1
logging==0.4.9.6